clean up; needs major refactoring for festim and dolfinx upgrades
