--- Simple Tube PAV Extractor/README.md
@@ -1,1 +1,0 @@
-This folder includes parametric studies for a simple tube Permeation Against Vacuum (PAV) Extractor to determine extractor the dependence of PAV efficiency on extractor parameters.
