+++ festim_scripts/README.md
@@ -1,0 +1,1 @@
+This folder includes parametric studies for a simple tube Permeation Against Vacuum (PAV) Extractor to determine extractor the dependence of PAV efficiency on extractor parameters.
